+++ src/main/typescript/rug/operations/Handlers.ts
@@ -274,11 +274,8 @@ export type MessageAddress = UserAddress | ChannelAddress;
  * Uncorrelated message to the bot
  */
 export class DirectedMessage extends LocallyRenderedMessage<"directed"> {
-  public kind: "directed" = "directed";
 
-  public id?: string;
-  // format has to be sortable like 2017-06-08T20:00:59.188Z or epoch
-  public timestamp?: string;
+  public kind: "directed" = "directed";
 
   constructor(body: string, address: MessageAddress, contentType?: MessageMimeType) {
     super();
@@ -287,6 +284,20 @@ export class DirectedMessage extends LocallyRenderedMessage<"directed"> {
       this.contentType = contentType;
     }
     this.addAddress(address);
+  }
+}
+
+/**
+ * Message that can be re-written in the bot
+ */
+export class UpdatableMessage extends DirectedMessage {
+
+  public id: string;
+  public timestamp: string = new Date().getTime().toString();
+
+  constructor(id: string, body: string, address: MessageAddress, contentType?: MessageMimeType) {
+    super(body, address, contentType);
+    this.id = id;
   }
 }